Changed KIC target
The original example used a KIC target that seems to be an eclipsing binary. As such, the red_chi2 value drifted a lot over the course of a single quarter.

I've switched to Kepler-78 instead, and the red_chi2, though not, fixed, varies by less than 2% over the course of one quarter. Unfortunately, the lightcurve data doesn't include a Kepler magnitude for some reason. So we'll have to be careful about plotting red_chi2 vs. KepMag.